Vognkontroller: working on change from double to int, added use Z-button to clear trip counter, and added way-points X and Y.

+++ Vognkontroller/src/My_filelogger.cpp
@@ -25,7 +25,7 @@ void My_filelogger::initialize() {
     logg('X', 0.0);
 }
 
-void My_filelogger::logg(char key, double data) {
+void My_filelogger::logg(char key, int data) {
     auto timestamp = std::chrono::duration_cast<std::chrono::milliseconds>(
             std::chrono::system_clock::now().time_since_epoch()) - the_zero_time;
     std::lock_guard<std::mutex> guard(the_mutex);
